+++ app/Http/Controllers/Admin/Gostaresh/GDPPartController.php
@@ -4,10 +4,10 @@ namespace App\Http\Controllers\Admin\Gostaresh;
 
 use App\Exports\Gostaresh\GDPPart\ListExport;
 use App\Http\Controllers\Controller;
-use App\Models\Index\GDPPart;
-use Illuminate\Http\Request;
-use Illuminate\Support\Facades\Auth;
 use App\Http\Requests\Gostaresh\GDPPart\GDPPartRequest;
+use App\Models\Index\GDPPart;
+use Illuminate\Http\Response;
+use Illuminate\Support\Facades\Auth;
 use Maatwebsite\Excel\Facades\Excel;
 use PDF;
 
@@ -17,10 +17,12 @@ class GDPPartController extends Controller
     /**
      * Display a listing of the resource.
      *
-     * @return \Illuminate\Http\Response
+     * @return Response
      */
     public function index()
     {
+        $this->authorize("view-any-GDPPart");
+
         $query = GDPPart::whereRequestsQuery();
 
         $filterColumnsCheckBoxes = GDPPart::$filterColumnsCheckBoxes;
@@ -63,30 +65,33 @@ class GDPPartController extends Controller
     /**
      * Show the form for creating a new resource.
      *
-     * @return \Illuminate\Http\Response
+     * @return Response
      */
     public function create()
     {
+        $this->authorize("create-any-GDPPart");
+
         return view('admin.gostaresh.gdp-part.create.create');
     }
 
     /**
      * Store a newly created resource in storage.
      *
-     * @param  GDPPartRequest  $request
-     * @return \Illuminate\Http\Response
+     * @param GDPPartRequest $request
+     * @return Response
      */
     public function store(GDPPartRequest $request)
     {
-        GDPPart::create(array_merge(['user_id' => Auth::id()] , $request->validated()));
+        $this->authorize("create-any-GDPPart");
+        GDPPart::create(array_merge(['user_id' => Auth::id()], $request->validated()));
         return back()->with('success', __('titles.success_store'));
     }
 
     /**
      * Display the specified resource.
      *
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
+     * @param int $id
+     * @return Response
      */
     public function show(GDPPart $gdpPart)
     {
@@ -96,35 +101,41 @@ class GDPPartController extends Controller
     /**
      * Show the form for editing the specified resource.
      *
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
+     * @param int $id
+     * @return Response
      */
     public function edit(GDPPart $gdpPart)
     {
-        return view('admin.gostaresh.gdp-part.edit.edit', compact('gdpPart'));
+        $this->authorize("edit-any-GDPPart");
+
+        return view(
+            'admin.gostaresh.gdp-part.edit.edit', compact('gdpPart'));
     }
 
     /**
      * Update the specified resource in storage.
      *
-     * @param  GDPPartRequest  $request
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
+     * @param GDPPartRequest $request
+     * @param int $id
+     * @return Response
      */
     public function update(GDPPartRequest $request, GDPPart $gdpPart)
     {
+        $this->authorize("edit-any-GDPPart");
         $gdpPart->update($request->validated());
+
         return back()->with('success', __('titles.success_update'));
     }
 
     /**
      * Remove the specified resource from storage.
      *
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
+     * @param int $id
+     * @return Response
      */
     public function destroy(GDPPart $gdpPart)
     {
+        $this->authorize("delete-any-GDPPart");
         $gdpPart->delete();
         return back()->with('success', __('titles.success_delete'));
     }
